refactor: 관리자 기능 코드 수정
